NEW LocationManager and LocCalculate class ,MyLocationListener DELETED
